UiActions moves to semantic-ui
It was never a mindconnect concern: it takes a controller method and
returns a UiTrigger, both of which are the framework's vocabulary. Sitting
in chatui.ui only made admin-ui depend on the chat module for it — the
wart the first commit named.

It now lives in ai.mindconnect.ui.mvc in mc-semantic-ui-core, beside the
ssr package, with Spring MVC staying an optional dependency so the model
and the JavaFX renderers are untouched. Six tests came along.

NOTE FOR MERGING: this pins semantic-ui.version to 0.2.2-SNAPSHOT, so the
branch needs the local install until mc-semantic-ui cuts 0.2.2. Replace
the pin with the release before merging; nothing else here depends on the
snapshot.
