+++ agents/server/mc-agent-admin-ui-rest/src/main/java/ai/mindconnect/adminui/ui/component/ToolDetailComponent.java
@@ -4,8 +4,8 @@ import ai.mindconnect.chatui.ui.UiComponent;
 import ai.mindconnect.agent.domain.AgentDefinition;
 import ai.mindconnect.adminui.ui.controller.AgentUiController;
 
-import static ai.mindconnect.chatui.ui.UiActions.ROW_ID;
-import static ai.mindconnect.chatui.ui.UiActions.trigger;
+import static ai.mindconnect.ui.mvc.UiActions.ROW_ID;
+import static ai.mindconnect.ui.mvc.UiActions.trigger;
 import static org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder.on;
 import ai.mindconnect.agent.tool.AgentTool;
 import ai.mindconnect.agent.tool.ToolRegistry;
