+++ agents/server/mc-agent-chat-ui-rest/src/main/java/ai/mindconnect/chatui/ui/component/ApprovalCardComponent.java
@@ -5,7 +5,7 @@ import ai.mindconnect.ui.model.UiAction;
 import ai.mindconnect.ui.model.UiTrigger;
 import ai.mindconnect.chatui.ui.controller.ChatUiController;
 
-import static ai.mindconnect.chatui.ui.UiActions.trigger;
+import static ai.mindconnect.ui.mvc.UiActions.trigger;
 import static org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder.on;
 import ai.mindconnect.ui.model.UiList;
